+++ src/components/pages/Home.jsx
@@ -1,5 +1,4 @@
-function Home() {
+export const Home = () => {
     return <h1>Welcome to the Home Page!</h1>;
   }
   
-  export default Home;
