+++ src/services/recipeServices.js
@@ -1,0 +1,27 @@
+export const fetchRecipes = async (setRecipes, token) => {
+  fetch("http://localhost:8000/recipes", {
+    headers: {
+      Authorization: `Token ${token}`,
+      "Content-Type": "application/json",
+    },
+  }).then(async (response) => {
+    const data = await response.json();
+    setRecipes(data);
+  });
+};
+
+export async function createRecipe(postData) {
+  const token = localStorage.getItem("auth_token");
+
+  const formData = new FormData();
+  Object.keys(postData).forEach((key) => formData.append(key, postData[key]));
+
+  const response = await fetch("http://localhost:8000/recipes", {
+    method: "POST",
+    headers: {
+      Authorization: `Token ${token}`,
+    },
+    body: formData,
+  });
+  return await response.json();
+}
